Move to_utf8 method to helpers
